+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 * core: Converter will force initialize companion objects to better support `init { Converter.use {...} }`
+* core: Cache keepAlive function introduced to prolong external cache entries
 
 # 1.6.5
 * core: simple Cache with expiration timer implemented